requirements.txt
torch
torchaudio
sentencepiece
pyyaml
wandb
soundfile
numpy
k2
psutil
thop
openai-whisper
datasets